remove ReentrantImmediateEventExecutor
