--- common/src/main/java/io/netty/util/concurrent/ReentrantImmediateEventExecutor.java
@@ -1,82 +1,0 @@
-/*
- * Copyright 2016 The Netty Project
- *
- * The Netty Project licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package io.netty.util.concurrent;
-
-import io.netty.util.internal.logging.InternalLogger;
-import io.netty.util.internal.logging.InternalLoggerFactory;
-
-import java.util.ArrayDeque;
-import java.util.Queue;
-
-/**
- * Executes {@link Runnable} objects in the caller's thread. If the {@link #execute(Runnable)} is reentrant it will be
- * queued until the original {@link Runnable} finishes execution.
- * <p>
- * All {@link Throwable} objects thrown from {@link #execute(Runnable)} will be swallowed and logged. This is to ensure
- * that all queued {@link Runnable} objects have the chance to be run.
- */
-public final class ReentrantImmediateEventExecutor extends ImmediateEventExecutor {
-    public static final ReentrantImmediateEventExecutor INSTANCE = new ReentrantImmediateEventExecutor();
-    private static final InternalLogger logger = InternalLoggerFactory.getInstance(ImmediateEventExecutor.class);
-    /**
-     * A Runnable will be queued if we are executing a Runnable. This is to prevent a {@link StackOverflowError}.
-     */
-    private static final FastThreadLocal<Queue<Runnable>> DELAYED_RUNNABLES = new FastThreadLocal<Queue<Runnable>>() {
-        @Override
-        protected Queue<Runnable> initialValue() throws Exception {
-            return new ArrayDeque<Runnable>();
-        }
-    };
-    /**
-     * Set to {@code true} if we are executing a runnable.
-     */
-    private static final FastThreadLocal<Boolean> RUNNING = new FastThreadLocal<Boolean>() {
-        @Override
-        protected Boolean initialValue() throws Exception {
-            return false;
-        }
-    };
-
-    private ReentrantImmediateEventExecutor() { }
-
-    @Override
-    public void execute(Runnable command) {
-        if (command == null) {
-            throw new NullPointerException("command");
-        }
-        if (!RUNNING.get()) {
-            RUNNING.set(true);
-            try {
-                command.run();
-            } catch (Throwable cause) {
-                logger.info("Throwable caught while executing Runnable {}", command, cause);
-            } finally {
-                Queue<Runnable> delayedRunnables = DELAYED_RUNNABLES.get();
-                Runnable runnable;
-                while ((runnable = delayedRunnables.poll()) != null) {
-                    try {
-                        runnable.run();
-                    } catch (Throwable cause) {
-                        logger.info("Throwable caught while executing Runnable {}", runnable, cause);
-                    }
-                }
-                RUNNING.set(false);
-            }
-        } else {
-            DELAYED_RUNNABLES.get().add(command);
-        }
-    }
-}
